Update HookedProxy.py
Refactored the Hook class. On further consideration, the idea of implementing callback functions was confusing. I decided it was better to have the user just extend the Hook class and override onHandleRequest and onHandleResponse. I removed the callbacks dict since these functions can now be called directly by the user. State can still be used to pass information between the request and response handler. I'm debating if it's cleaner to just take the active part out of the state dict, but that will get fixed later if I decide to