Simplified code
Removed some custom error types in favor of built in python ones. Still need to remove InvalidHookStateError. I think it's better to take 'active' out of state and make it a class variable and just leave state for the user. Forcing them to keep active in the state is more complicated and doesn't mimic the javascript design patter well since JS maintains a dict when you add to it. Ours overrides by default.